New structure maplayers and fixed z-index
