--- src/components/KadastraleGrens.vue
@@ -1,11 +1,0 @@
-<template>
-  <div class="hidden">
-  </div>
-</template>
-<script src="./kadastrale-grens.js"></script>
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped>
-  .hidden {
-    display: none;
-  }
-</style>
