--- src/components/SatelliteImages.vue
@@ -1,11 +1,0 @@
-<template>
-  <div class="hidden">
-  </div>
-</template>
-<script src="./satellite-images.js"></script>
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped>
-  .hidden {
-    display: none;
-  }
-</style>
