+++ src/components/VMapLayers.vue
@@ -1,0 +1,6 @@
+<template>
+</template>
+
+<script src="./v-map-layers.js"></script>
+
+<style></style>
