--- src/components/VStreamlines.vue
@@ -1,11 +1,0 @@
-<template>
-  <div class="hidden">
-  </div>
-</template>
-<script src="./v-streamlines.js"></script>
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped>
-  .hidden {
-    display: none;
-  }
-</style>
