+++ src/components/v-map-layers.js
@@ -1,0 +1,35 @@
+import {
+  bus
+} from '@/event-bus.js';
+import {
+  mapLayers
+} from './map-layers-config.js'
+import {
+  getGeeSource
+} from './getGeeSource.js'
+export default {
+  name: 'v-map-layers',
+  data() {
+    return {
+      map: null,
+      dateBegin: "2016-07-20",
+      dateEnd: "2016-07-21",
+      bands: ["red", "green", "blue"],
+    }
+  },
+  methods: {
+    deferredMountedTo(map) {
+      _.each(mapLayers, (layer) => {
+
+        if (layer.layertype === 'mapbox-layer') {
+          _.each(layer.data, (maplayer) => {
+            map.addLayer(maplayer);
+          })
+          bus.$emit('add-layer', layer);
+        } else if (layer.layertype === 'gee-layer') {
+          getGeeSource(map, layer, this.dateBegin, this.dateEnd, this.bands)
+        }
+      })
+    }
+  }
+}
